orchestrators/ticket-purchase-orchestrator/requirements.txt
flask==3.1.3
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.8.0
gunicorn==22.0.0

# Testing
pytest==8.3.5
pytest-mock==3.14.0
grpcio==1.78.0
grpcio-tools==1.78.0
pika==1.3.2
